+++ configurator_test.go
@@ -51,29 +51,6 @@ func TestConfigurator_Load_IgnoreUnexportedField(t *testing.T) {
 	assert.Equal(t, expected, actual)
 }
 
-func TestConfigurator_Load_IgnoreStructs(t *testing.T) {
-	type subconfig struct {
-		value string `default:"default"`
-	}
-
-	type config struct {
-		sconfig subconfig
-	}
-
-	expected := config{
-		sconfig: subconfig{
-			value: "value",
-		},
-	}
-	actual := expected
-
-	configurator := nest.NewConfigurator()
-
-	err := configurator.Load(&actual)
-	require.NoError(t, err)
-	assert.Equal(t, expected, actual)
-}
-
 func TestConfigurator_Load_Ignored(t *testing.T) {
 	type config struct {
 		Value string `ignored:"true" default:"default"`
@@ -413,6 +390,58 @@ func TestConfigurator_Load_Default(t *testing.T) {
 	assert.Equal(t, expected, actual)
 }
 
+func TestConfigurator_Load_Struct(t *testing.T) {
+	type subconfig struct {
+		Value string `default:"default"`
+	}
+
+	type config struct {
+		Sconfig subconfig
+	}
+
+	expected := config{
+		Sconfig: subconfig{
+			Value: "default",
+		},
+	}
+	actual := config{}
+
+	configurator := nest.NewConfigurator()
+
+	err := configurator.Load(&actual)
+	require.NoError(t, err)
+	assert.Equal(t, expected, actual)
+}
+
+func TestConfigurator_Load_StructEnvWithPrefix(t *testing.T) {
+	type subconfig struct {
+		Value string `env:""`
+	}
+
+	type config struct {
+		Sconfig subconfig
+	}
+
+	expected := config{
+		Sconfig: subconfig{
+			Value: "value",
+		},
+	}
+	actual := config{}
+
+	configurator := nest.NewConfigurator()
+	configurator.SetEnvPrefix("app")
+
+	os.Clearenv()
+	os.Setenv("APP_SCONFIG_VALUE", "value")
+
+	err := configurator.Load(&actual)
+	require.NoError(t, err)
+	assert.Equal(t, expected, actual)
+
+	os.Clearenv()
+}
+
 func TestConfigurator_Load_Types(t *testing.T) {
 	type config struct {
 		String string
@@ -595,8 +624,8 @@ func TestConfigurator_Load_PrecedenceOrder(t *testing.T) {
 
 	expected := config{
 		Override: "override",
-		Flag:     "value",
-		Env:      "value",
+		Flag:     "flag",
+		Env:      "env",
 		Default:  "default",
 	}
 	actual := config{
@@ -604,10 +633,10 @@ func TestConfigurator_Load_PrecedenceOrder(t *testing.T) {
 	}
 
 	configurator := nest.NewConfigurator()
-	configurator.SetArgs([]string{"program", "--flag", "value"})
+	configurator.SetArgs([]string{"program", "--flag", "flag"})
 
 	os.Clearenv()
-	os.Setenv("ENV", "value")
+	os.Setenv("ENV", "env")
 
 	err := configurator.Load(&actual)
 	require.NoError(t, err)