+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Missing global `SetName` function
 - `SetArgs` to set command line arguments manually
 - Add struct tag constant values
+- Support child structs
+
+### Changes
+
+- Refactor internals
 
 
 ## [0.2.0] - 2017-12-19